infrastructures/config/dev/requirements.txt
fastapi[all]
SQLAlchemy==1.4.48
asyncpg==0.27.0
aiofiles==23.1.0
orjson==3.8.8
httpx==0.23.3
pydantic==1.10.7
passlib==1.7.4
python-dotenv==1.0.0
python-jose==3.3.0